Move files to sub-package mac
